// functions/src/types.ts
export enum Collections {
  Users = "Users",
  Yaytsos = "YAYTSOS",
  TxLogs = "TxLogs",
  // NFTS = "NFTS",
  NFTS = "NFTS_POLY",

  Guestlist = "Guestlist",
  // Cartons = "CARTONS",
  Cartons = "CARTONS_POLY",

  // CartonClaims = "CartonClaims",
  CartonClaims = "CartonClaims_POLY",
  Grid = "Grid",
}